+++ PagaTapp/PagaTapp/Modules/BanksList/BanksListPresenter.swift
@@ -6,13 +6,6 @@
 //
 
 import Foundation
-protocol BanksListPresenterProtocol {
-    var view: BanksListViewControllerProtocol? {get set}
-    var router: BanksListRouterProtocol? {get set}
-    var interactor: BanksListInteractorProtocol? {get set}
-    
-    func listBanks()
-}
 class BanksListPresenter: BanksListPresenterProtocol {
     func listBanks() {
         self.interactor?.loadBanks { result in
